m fix

+++ force_calc.f90
@@ -12,7 +12,7 @@ subroutine full_bead_forces
 
  do bat = 0, Nbatches - 1 !batch index
 
-	if ((pot_model .eq. 6).and.(mod(t,2000).eq.0)) then 
+	if ((pot_model .eq. 6).and.(mod(t,2000).eq.0).and.(t.gt.0)) then 
 		call system("rm "//trim(sys_label)//trim(str(pid))//".out") !clean out .out files
 	endif
